requirements.txt
requests
pysimplegui
pandas
python-telegram-bot
web3
py-etherscan-api
Pillow